server/requirements.txt
fastapi
uvicorn
moviepy
google-genai
python-dotenv
python-multipart
gTTS
